pyproject.toml
[tool.poetry]
name = "data-analysis-convai"
version = "0.1.0"
description = "Conversation chatbot that translates user query into SQL to fetch the data from a database and present it in a readable form."
authors = ["janis cakstins"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.5"
python-dotenv = "^1.0.1"
plotly = "^5.21.0"
psycopg2-binary = "^2.9.9"
chainlit = "^1.2"
asyncio = "^3.4"
httpx = "^0.27"
pydantic = "2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
